requirements.txt
Flask==3.0.2
flask-restx==1.3.0
PyJWT==2.8.0
bcrypt==4.0.1
Faker==18.0.0
dnspython==2.3.0
requests==2.28.1
beautifulsoup4==4.11.2
